roll-stock-service/requirements.txt
uvicorn~=0.30.1
fastapi~=0.104.1
python-dotenv~=1.0.0
SQLAlchemy~=2.0.22
pydantic~=2.4.2
h11~=0.14.0
click~=8.1.7
Mako~=1.2.4
MarkupSafe~=2.1.3
anyio~=3.7.1
sniffio~=1.3.0
alembic~=1.12.1
starlette~=0.27.0
greenlet~=3.0.1
psycopg2~=2.9.9
fastapi_sqlalchemy~=0.2.1
aiohttp~=3.8.6